Внесены изменения в готовый бэк по чек-листу
